requirements.txt
annotated-types==0.6.0
anyio==3.7.1
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
distro==1.8.0
entrypoints==0.3
exceptiongroup==1.2.0
flake8==3.7.9
Flask==3.0.0
Flask-API==2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mccabe==0.6.1
openai==1.3.7
pycodestyle==2.5.0
pydantic==2.5.2
pydantic_core==2.14.5
pyflakes==2.1.1
python-dotenv==1.0.0
regex==2023.10.3
requests==2.31.0
sniffio==1.3.0
tiktoken==0.5.2
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.1.0
Werkzeug==3.0.1
